Rework the plugin mecanism
